add TestBase and refactor Tests
